requirements.txt
# ============================================================================
# FUEL COPILOT BACKEND - REQUIREMENTS
# ============================================================================
# Version: 3.10.8
# Platform: Windows Azure VM + Linux
# Date: December 2025

# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data validation
pydantic==2.4.2

# WebSockets
websockets==12.0

# Database
pymysql==1.1.0
PyMySQL[rsa]==1.1.0
sqlalchemy==2.0.23

# Data processing
pandas>=2.1.0
numpy>=1.26.0

# Environment variables
python-dotenv==1.0.0

# HTTP client
httpx==0.25.0

# Authentication
PyJWT==2.8.0

# YAML config
PyYAML==6.0.1

# Prometheus metrics (optional)
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# Redis cache (optional)
redis==5.0.1
